requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
numpy
scikit-learn
faiss-cpu
annoy
sqlalchemy
